reorganized repo layout in a vareity of ways: - moved webhooks along side api/; instead of nested - renamed subhub_api.yaml -> swagger.yaml - combined exceptions test files - moved test_app under unit/ - renamed auth_validation -> authentication - renamed subhub_dynamodb -> db - renamed p -> pipeline - renamed payment_intents -> intents - renamed behave/ -> bdd/
