+++ subhub/webhooks/routes/pipeline.py
@@ -5,9 +5,9 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-from subhub.api.webhooks.routes.firefox import FirefoxRoute
-from subhub.api.webhooks.routes.salesforce import SalesforceRoute
-from subhub.api.webhooks.routes.static import StaticRoutes
+from subhub.webhooks.routes.firefox import FirefoxRoute
+from subhub.webhooks.routes.salesforce import SalesforceRoute
+from subhub.webhooks.routes.static import StaticRoutes
 
 
 class RoutesPipeline:
