+++ subhub/webhooks/stripe/customer.py
@@ -10,8 +10,8 @@ import time
 
 import stripe
 
-from subhub.api.webhooks.stripe.abstract import AbstractStripeWebhookEvent
-from subhub.api.webhooks.routes.static import StaticRoutes
+from subhub.webhooks.stripe.abstract import AbstractStripeWebhookEvent
+from subhub.webhooks.routes.static import StaticRoutes
 from subhub.exceptions import ClientError
 
 from subhub.log import get_logger
