+++ subhub/webhooks/stripe/invoices.py
@@ -7,8 +7,8 @@
 
 import json
 
-from subhub.api.webhooks.stripe.abstract import AbstractStripeWebhookEvent
-from subhub.api.webhooks.routes.static import StaticRoutes
+from subhub.webhooks.stripe.abstract import AbstractStripeWebhookEvent
+from subhub.webhooks.routes.static import StaticRoutes
 
 from subhub.log import get_logger
